+++ src/app/notifications/notification-page-client.tsx
@@ -4,7 +4,7 @@ import React, { useState } from "react";
 import { Settings, RefreshCw, CheckCheck } from "lucide-react";
 import Link from "next/link";
 import { useNotifications } from "@/hooks/use-notifications";
-import { NotificationList } from "@/components/notification/NotificationList";
+import { NotificationList } from "@/components/notification/notification-list";
 import { Button } from "@/components/ui/button";
 import { Tabs, TabsContent, TabsList, TabsTrigger } from "@/components/ui/tabs";
 import {
