+++ src/components/ganttv3/inline-task-edit-panel.tsx
