+++ src/components/ganttv3/task-bar.tsx
