+++ src/components/ganttv3/view-switcher.tsx
